chore(#8): extend unit tests
