+++ src/interfaces/IButton.interface.tsx
@@ -1,0 +1,4 @@
+export interface IButton {
+  type: 'skip' | 'done' | 'next' | 'prev';
+  label?: string;
+}
