--- src/interfaces/IDoneButton.interface.tsx
@@ -1,3 +1,0 @@
-export interface IDoneButton {
-  doneLabel?: string;
-}
