--- src/interfaces/INextButton.interface.tsx
@@ -1,3 +1,0 @@
-export interface INextButton {
-  nextLabel?: string;
-}
