--- src/interfaces/ISkipButton.interface.tsx
@@ -1,3 +1,0 @@
-export interface ISkipButton {
-  skipLabel?: string;
-}
