+++ src/interfaces/index.d.tsx
@@ -1,0 +1,9 @@
+export * from './IAnimatedValues.interface';
+export * from './IButton.interface';
+export * from './IDotContainer.interface';
+export * from './IItem.interface';
+export * from './INextContainer.interface';
+export * from './IPrevContainer.interface';
+export * from './ISlide.interface';
+export * from './ISliderIntro.interface';
+export * from './IStatusBar.interface';
